Implement feedback from PR comments
