+++ gizmos/angle.py
@@ -9,10 +9,7 @@ from ..utilities.constants import QUARTER_TURN
 from ..utilities.draw import coords_arc_2d
 from ..utilities.math import pol2cart
 from ..utilities.view import get_scale_from_pos
-from .base import (
-    ConstraintGenericGGT,
-    ConstraintGizmoGeneric,
-)
+from .base import ConstraintGenericGGT, ConstraintGizmoGeneric
 from .utilities import draw_arrow_shape, get_arrow_size, get_overshoot
