+++ gizmos/base.py
@@ -1,51 +1,7 @@
-from enum import Enum, auto
-
 from .. import global_data
 from ..declarations import Operators
 from ..model.types import GenericConstraint
-from ..utilities.preferences import get_prefs
-
-
-class Color(Enum):
-    Default = auto()
-    Failed = auto()
-    Reference = auto()
-    Text = auto()
-
-
-def get_constraint_color_type(constraint: GenericConstraint):
-    if constraint.failed:
-        return Color.Failed
-    if constraint.is_reference:
-        return Color.Reference
-    return Color.Default
-
-
-def get_color(color_type: Color, highlit: bool):
-    c_theme = get_prefs().theme_settings.constraint
-    theme_match = {
-        # (type, highlit): color
-        (Color.Default, False): c_theme.default,
-        (Color.Default, True): c_theme.highlight,
-        (Color.Failed, False): c_theme.failed,
-        (Color.Failed, True): c_theme.failed_highlight,
-        (Color.Reference, False): c_theme.reference,
-        (Color.Reference, True): c_theme.reference_highlight,
-        (Color.Text, False): c_theme.text,
-        (Color.Text, True): c_theme.text_highlight,
-    }
-    return theme_match[(color_type, highlit)]
-
-
-def set_gizmo_colors(gz, constraint):
-    color_type = get_constraint_color_type(constraint)
-    color = get_color(color_type, highlit=False)
-    color_highlight = get_color(color_type, highlit=True)
-
-    gz.color = color[0:-1]
-    gz.alpha = color[-1]
-    gz.color_highlight = color_highlight[0:-1]
-    gz.alpha_highlight = color_highlight[-1]
+from .utilities import get_color, get_constraint_color_type, set_gizmo_colors
 
 
 class ConstraintGizmo:
